refactor: fix imports due to moving types into rel
apache/gravitino#8589
