+++ clients/client-python/gravitino/api/rel/types/json_serdes/base.py
@@ -24,7 +24,7 @@ from gravitino.api.expressions.distributions.distribution import Distribution
 from gravitino.api.expressions.expression import Expression
 from gravitino.api.expressions.sorts.sort_order import SortOrder
 from gravitino.api.rel.indexes.index import Index
-from gravitino.api.types.types import Type
+from gravitino.api.rel.types.types import Type
 from gravitino.dto.rel.partitioning.partitioning import Partitioning
 from gravitino.dto.rel.partitions.partition_dto import PartitionDTO
