+++ clients/client-python/gravitino/dto/rel/column_dto.py
@@ -24,9 +24,9 @@ from dataclasses_json import DataClassJsonMixin, config
 
 from gravitino.api.expressions.expression import Expression
 from gravitino.api.rel.column import Column
-from gravitino.api.types.json_serdes.type_serdes import TypeSerdes
-from gravitino.api.types.type import Type
-from gravitino.api.types.types import Types
+from gravitino.api.rel.types.json_serdes.type_serdes import TypeSerdes
+from gravitino.api.rel.types.type import Type
+from gravitino.api.rel.types.types import Types
 from gravitino.dto.rel.expressions.json_serdes.column_default_value_serdes import (
     ColumnDefaultValueSerdes,
 )
