+++ clients/client-python/gravitino/dto/rel/json_serdes/distribution_serdes.py
@@ -19,7 +19,7 @@
 from typing import Any
 
 from gravitino.api.expressions.distributions.strategy import Strategy
-from gravitino.api.types.json_serdes.base import JsonSerializable
+from gravitino.api.rel.types.json_serdes.base import JsonSerializable
 from gravitino.dto.rel.distribution_dto import DistributionDTO
 from gravitino.dto.rel.expressions.json_serdes._helper.serdes_utils import SerdesUtils
 from gravitino.utils.precondition import Precondition
