+++ clients/client-python/gravitino/dto/rel/partitions/json_serdes/partition_dto_serdes.py
@@ -17,7 +17,7 @@
 
 from typing import Any, Dict
 
-from gravitino.api.types.json_serdes.base import JsonSerializable
+from gravitino.api.rel.types.json_serdes.base import JsonSerializable
 from gravitino.dto.rel.partitions.json_serdes._helper.serdes_utils import SerdesUtils
 from gravitino.dto.rel.partitions.partition_dto import PartitionDTO
